feat: refactor enums to new Core.Enums namespace
Changed the namespace for `ERegistrationMode` and `ERegistrationStrategyType` enums from `OnForkHub.Application.DependencyInjection` to `OnForkHub.Core.Enums`. This restructuring improves organization without altering the enum members.
